Antigravity-AgenticAIs copy/Agent-3/requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.19
python-dotenv>=1.0.0
